foundation/foundation-lineage/aissemble-foundation-model-lineage/pyproject.toml
[tool.poetry]
name = "aissemble-foundation-model-lineage"
version = "1.7.0.dev"
authors = ["aiSSEMBLE Baseline Community <aissemble@bah.com>"]
description = "Model lineage API"
readme = "README.md"
packages = [
    {include = "aissemble_model_lineage", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.11.4, <4"
openlineage-python = {version = "^0.21.1"}

#[tool.poetry.group.monorepo.dependencies]
aissemble-foundation-data-lineage-python = { path = "../foundation-data-lineage/aissemble-foundation-data-lineage-python", develop = true}

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
behave = "^1.2.6"
nose = ">=1.3.7"
kappa-maki = ">=1.0.1"




[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"